+++ src/descriptors/_element_center_test.js
@@ -9,14 +9,10 @@ var Position = require("../values/position.js");
 
 describe("ElementCenter", function() {
 
-	var TOP = 10;
-	var RIGHT = 150;
-	var BOTTOM = 70;
-	var LEFT = 20;
-
 	var frame;
 	var element;
 	var center;
+	var middle;
 
 	before(function(done) {
 		frame = quixote.createFrame(500, 500, { stylesheet: "/base/src/__reset.css" }, done);
@@ -33,7 +29,8 @@ describe("ElementCenter", function() {
 			"<p id='one' style='position: absolute; left: 20px; width: 130px; top: 10px; height: 60px'>one</p>"
 		);
 		element = frame.getElement("#one");
-		center = new ElementCenter(element);
+		center = ElementCenter.x(element);
+		middle = ElementCenter.y(element);
 	});
 
 	it("is a descriptor", function() {
@@ -41,16 +38,20 @@ describe("ElementCenter", function() {
 	});
 
 	it("resolves to value", function() {
-		assert.objEqual(center.value(), Position.x(85));
+		assert.objEqual(center.value(), Position.x(85), "center");
+		assert.objEqual(middle.value(), Position.y(40), "middle");
 	});
 
 	it("converts comparison arguments", function() {
-		assert.objEqual(center.convert(13), Position.x(13), "should convert numbers to positions");
+		assert.objEqual(center.convert(13), Position.x(13), "should convert numbers to x-positions");
+		assert.objEqual(middle.convert(13), Position.y(13), "should convert numbers to y-positions");
+
 		assert.equal(center.convert(center), center, "should return descriptors as-is");
 	});
 
 	it("converts to string", function() {
-		assert.equal(center.toString(), "center of element '" + element.description() + "'");
+		assert.equal(center.toString(), "center of element '" + element.description() + "'", "center");
+		assert.equal(middle.toString(), "middle of element '" + element.description() + "'", "middle");
 	});
 
 });