+++ src/descriptors/element_center.js
@@ -5,31 +5,45 @@ var ensure = require("../util/ensure.js");
 var Descriptor = require("./descriptor.js");
 var Position = require("../values/position.js");
 
-var Me = module.exports = function ElementCenter(element) {
-	// TODO: ensure.signature (circular dependency)
+var X_DIMENSION = "x";
+var Y_DIMENSION = "y";
 
+var Me = module.exports = function ElementCenter(dimension, element) {
+	// TODO: ensure.signature (circular dependency)
+	ensure.that(dimension === X_DIMENSION || dimension === Y_DIMENSION, "Unrecognized dimension: " + dimension);
+
+	this._dimension = dimension;
 	this._element = element;
 };
 Descriptor.extend(Me);
 
 Me.x = function(element) {
-	return new Me(element);
+	return new Me(X_DIMENSION, element);
+};
+
+Me.y = function(element) {
+	return new Me(Y_DIMENSION, element);
 };
 
 Me.prototype.value = function value() {
 	ensure.signature(arguments, []);
 
 	var position = this._element.getRawPosition();
-	return Position.x(position.left + ((position.right - position.left) / 2));
+
+	if (this._dimension === X_DIMENSION) return Position.x(position.left + ((position.right - position.left) / 2));
+	else return Position.y(position.top + ((position.bottom - position.top) / 2));
 };
 
 Me.prototype.convert = function convert(arg) {
 	ensure.signature(arguments, [ [Number, Descriptor ]]);
 
-	if (typeof arg === "number") return Position.x(arg);
-	else return arg;
+	if (typeof arg !== "number") return arg;
+
+	if (this._dimension === X_DIMENSION) return Position.x(arg);
+	else return Position.y(arg);
 };
 
 Me.prototype.toString = function toString() {
-	return "center of element '" + this._element.description() + "'";
+	var description = (this._dimension === X_DIMENSION) ? "center" : "middle";
+	return description + " of element '" + this._element.description() + "'";
 };